reorganize, dependabot, test rbd
