+++ package.json
@@ -52,6 +52,7 @@
         "notistack": "^1.0.0",
         "nprogress": "^0.2.0",
         "react": "^16.12.0",
+        "react-beautiful-dnd": "^13.0.0",
         "react-dom": "^16.12.0",
         "react-markdown": "^4.3.1",
         "react-material-ui-carousel": "^1.8.3",
@@ -69,6 +70,7 @@
         "@types/node": "^12.12.55",
         "@types/nprogress": "^0.2.0",
         "@types/react": "^16.9.49",
+        "@types/react-beautiful-dnd": "^13.0.0",
         "@types/react-dom": "^16.9.4",
         "@types/react-swipeable-views": "^0.13.0",
         "@types/react-syntax-highlighter": "^11.0.5",
