+++ src/components/admin/AddMemberDialog.tsx
@@ -12,24 +12,19 @@ import { makeStyles, Theme } from "@material-ui/core/styles";
 import { Close } from "@material-ui/icons";
 import { Field, useFormikContext } from "formik";
 import { TextField } from "formik-material-ui";
-import { KeyboardDatePicker } from "formik-material-ui-pickers";
 import React, { useEffect, useState } from "react";
 import * as Yup from "yup";
-import { Event, eventTypeLabels } from "../interfaces";
-import { addEventSchema } from "../lib/validators";
-import ArrayField from "./ArrayField";
-import DynamicForm from "./DynamicForm";
-import TransitionSlide from "./TransitionSlide";
-import UploadImages from "./UploadImages";
+import { Member, memberPositionLabels } from "../../interfaces";
+import { addMemberSchema } from "../../lib/validators";
+import ArrayField from "../ArrayField";
+import DynamicForm from "../DynamicForm";
+import TransitionSlide from "../TransitionSlide";
+import UploadImage from "./UploadImage";
 
 interface Props {
     open: boolean;
-    initialValues?: Event;
-    addEvent?: (
-        event: Event,
-        images: { file: File | null; url: string }[],
-        progressCallback?: (progress: number) => any
-    ) => any;
+    initialValues?: Member;
+    addMember?: (member: Member, image?: File, progressCallback?: (progress: number) => any) => any;
     handleClose?: () => void;
 }
 
@@ -46,42 +41,37 @@ const useStyles = makeStyles((theme: Theme) => ({
     }
 }));
 
-const defaultInitialValues: Event = {
+const defaultInitialValues: Member = {
     id: "",
     name: "",
+    displayName: "",
     title: "",
-    term: "",
-    type: "otherEvent",
-    location: "",
+    email: "",
     description: "",
-    date: new Date().toISOString(),
-    photoCredits: [],
-    images: []
+    facts: [],
+    links: [],
+    positions: [],
+    rank: 0,
+    image: ""
 };
 
-const AddEventDialog: React.FC<Props> = ({ open, initialValues, addEvent, handleClose }: Props) => {
+const AddMemberDialog: React.FC<Props> = ({
+    open,
+    initialValues,
+    addMember,
+    handleClose
+}: Props) => {
     const classes = useStyles();
 
     const composedInitialValues = { ...defaultInitialValues, ...initialValues };
     const editing = !!composedInitialValues.id;
 
-    const [images, setImages] = useState<{ file: File | null; url: string }[]>([]);
+    const [image, setImage] = useState<{ file: File; url: string } | null>(null);
     const [uploading, setUploading] = useState<boolean>(false);
     const [uploadingProgress, setUploadingProgress] = useState<number>(0);
 
-    const handleImageUpload = (selectedFiles: FileList, previews: string[]) => {
-        const newImages = [
-            ...Array.from(selectedFiles).map((file, idx) => ({ file, url: previews[idx] }))
-        ];
-        setImages([...images, ...newImages]);
-    };
-
-    const handleClearImage = (idx: number) => {
-        setImages([...images.slice(0, idx), ...images.slice(idx + 1, images.length)]);
-    };
-
-    const handleClearImages = () => {
-        setImages([]);
+    const handleImageUpload = (selectedFile: File, preview: string) => {
+        setImage({ file: selectedFile, url: preview });
     };
 
     const imageUploadProgress = (progress: number) => {
@@ -93,10 +83,7 @@ const AddEventDialog: React.FC<Props> = ({ open, initialValues, addEvent, handle
     }, [uploading]);
 
     useEffect(() => {
-        if (!open) setImages([]);
-        else {
-            setImages([...composedInitialValues.images.map(url => ({ file: null, url }))]);
-        }
+        if (!open) setImage(null);
     }, [open]);
 
     return (
@@ -108,7 +95,7 @@ const AddEventDialog: React.FC<Props> = ({ open, initialValues, addEvent, handle
         >
             <DialogTitle id="form-dialog-title">
                 <Typography component="p" variant="h6">
-                    {editing ? "Edit Event" : "Add Event"}
+                    {editing ? "Edit Member" : "Add Member"}
                 </Typography>
                 {handleClose && (
                     <IconButton
@@ -125,29 +112,20 @@ const AddEventDialog: React.FC<Props> = ({ open, initialValues, addEvent, handle
                     className={classes.form}
                     validateOnChange={false}
                     validateOnBlur={true}
-                    validationSchema={addEventSchema}
+                    validationSchema={addMemberSchema}
                     initialValues={composedInitialValues}
-                    onSubmit={async (data: Event, { setSubmitting }) => {
-                        const newImages = images.some(image => !!image.file);
-
+                    onSubmit={async (data: Member, { setSubmitting }) => {
                         setSubmitting(true);
-                        newImages && setUploading(true);
-
-                        // date picker gives the date as a date object, however we store it as string
-                        const event = {
-                            ...data,
-                            date:
-                                typeof data.date !== "string"
-                                    ? (data.date as Date).toISOString()
-                                    : data.date
-                        };
+                        image && setUploading(true);
 
                         // handle submit
-
-                        addEvent && (await addEvent(event, [...images], imageUploadProgress));
+                        if (addMember) {
+                            if (image) await addMember(data, image.file, imageUploadProgress);
+                            else addMember(data);
+                        }
                         handleClose && handleClose();
 
-                        newImages && setUploading(false);
+                        image && setUploading(false);
                         setSubmitting(false);
                     }}
                     fields={[
@@ -168,9 +146,20 @@ const AddEventDialog: React.FC<Props> = ({ open, initialValues, addEvent, handle
                             props: {
                                 component: TextField,
                                 variant: "outlined",
+                                name: "displayName",
+                                label: "Display Name",
+                                fullWidth: true
+                            }
+                        },
+                        {
+                            component: Field,
+                            props: {
+                                component: TextField,
+                                variant: "outlined",
                                 name: "title",
                                 label: "Title",
-                                fullWidth: true
+                                fullWidth: true,
+                                required: true
                             }
                         },
                         {
@@ -178,34 +167,8 @@ const AddEventDialog: React.FC<Props> = ({ open, initialValues, addEvent, handle
                             props: {
                                 component: TextField,
                                 variant: "outlined",
-                                name: "term",
-                                label: "Term",
-                                fullWidth: true
-                            }
-                        },
-                        {
-                            component: Field,
-                            props: {
-                                component: TextField,
-                                select: true,
-                                variant: "outlined",
-                                name: "type",
-                                label: "Type",
-                                children: Object.entries(eventTypeLabels).map(([value, label]) => (
-                                    <MenuItem key={value} value={value}>
-                                        {label}
-                                    </MenuItem>
-                                )),
-                                fullWidth: true
-                            }
-                        },
-                        {
-                            component: Field,
-                            props: {
-                                component: TextField,
-                                variant: "outlined",
-                                name: "location",
-                                label: "Location",
+                                name: "email",
+                                label: "Email Address",
                                 fullWidth: true
                             }
                         },
@@ -223,23 +186,10 @@ const AddEventDialog: React.FC<Props> = ({ open, initialValues, addEvent, handle
                             }
                         },
                         {
-                            component: Field,
-                            props: {
-                                component: KeyboardDatePicker,
-                                variant: "inline",
-                                inputVariant: "outlined",
-                                name: "date",
-                                label: "Date",
-                                format: "MM/dd/yyyy",
-                                type: "string",
-                                fullWidth: true
-                            }
-                        },
-                        {
                             component: ArrayField,
                             props: {
-                                name: "photoCredits",
-                                addLabel: "Add Photo Credit",
+                                name: "facts",
+                                addLabel: "Add Fact",
                                 schema: Yup.array().of(
                                     Yup.object().default(() => ({
                                         props: {
@@ -247,23 +197,88 @@ const AddEventDialog: React.FC<Props> = ({ open, initialValues, addEvent, handle
                                             variant: "outlined",
                                             fullWidth: true
                                         },
-                                        fieldLabel: (idx: number) => `Photo Credit ${idx + 1}`,
+                                        fieldLabel: (idx: number) => `Fact ${idx + 1}`,
                                         initialValue: ""
                                     }))
                                 )
                             }
                         },
                         {
-                            component: UploadImages,
+                            component: ArrayField,
+                            props: {
+                                name: "links",
+                                addLabel: "Add Link",
+                                schema: Yup.array().of(
+                                    Yup.object().default({
+                                        title: Yup.object().default(() => ({
+                                            props: {
+                                                component: TextField,
+                                                variant: "outlined",
+                                                fullWidth: true
+                                            },
+                                            fieldLabel: (idx: number) => `Link Title ${idx + 1}`,
+                                            initialValue: ""
+                                        })),
+                                        link: Yup.object().default(() => ({
+                                            props: {
+                                                component: TextField,
+                                                variant: "outlined",
+                                                fullWidth: true
+                                            },
+                                            fieldLabel: (idx: number) => `Link ${idx + 1}`,
+                                            initialValue: ""
+                                        }))
+                                    })
+                                )
+                            }
+                        },
+                        {
+                            component: ArrayField,
+                            props: {
+                                name: "positions",
+                                addLabel: "Add Position",
+                                schema: Yup.array().of(
+                                    Yup.object().default(() => ({
+                                        props: {
+                                            component: TextField,
+                                            select: true,
+                                            variant: "outlined",
+                                            children: Object.entries(memberPositionLabels).map(
+                                                ([value, label]) => (
+                                                    <MenuItem key={value} value={value}>
+                                                        {label}
+                                                    </MenuItem>
+                                                )
+                                            ),
+                                            fullWidth: true
+                                        },
+                                        fieldLabel: (idx: number) => `Position ${idx + 1}`,
+                                        initialValue: "activeMember"
+                                    }))
+                                )
+                            }
+                        },
+                        {
+                            component: Field,
+                            props: {
+                                component: TextField,
+                                variant: "outlined",
+                                name: "rank",
+                                label: "Rank",
+                                type: "number",
+                                fullWidth: true
+                            }
+                        },
+                        {
+                            component: UploadImage,
                             props: {
                                 uploading,
-                                addLabel: "Add Image",
-                                fieldLabel: (idx: number) => `Image ${idx + 1}`,
+                                name: "image",
+                                label: "Image URL",
                                 uploadingProgress,
-                                images,
+                                image,
                                 onChange: handleImageUpload,
-                                clearImage: handleClearImage,
-                                clearImages: handleClearImages
+                                clearImage: () => setImage(null)
                             }
                         },
                         {
@@ -294,4 +309,4 @@ const AddEventDialog: React.FC<Props> = ({ open, initialValues, addEvent, handle
     );
 };
 
-export default AddEventDialog;
+export default AddMemberDialog;
