+++ src/components/admin/AdminLoading.tsx
@@ -1,7 +1,7 @@
 import { CircularProgress, Container } from "@material-ui/core";
 import { makeStyles, Theme } from "@material-ui/core/styles";
 import React from "react";
-import AdminLayout from "./layouts/AdminLayout";
+import AdminLayout from "../layouts/AdminLayout";
 
 const useStyles = makeStyles((theme: Theme) => ({
     paper: {
