+++ src/pages/admin/companies/index.tsx
@@ -3,8 +3,8 @@ import { GetStaticProps, NextPage } from "next";
 import { useSnackbar } from "notistack";
 import { useState } from "react";
 import useSWR from "swr";
-import AddCompanyDialog from "../../../components/AddCompanyDialog";
-import CompanyList from "../../../components/CompanyList";
+import AddCompanyDialog from "../../../components/admin/AddCompanyDialog";
+import CompanyList from "../../../components/admin/CompanyList";
 import { useConfirm } from "../../../components/ConfirmProvider";
 import AdminLayout from "../../../components/layouts/AdminLayout";
 import { Company, hasPermission, Member, User } from "../../../interfaces";
