+++ src/pages/admin/events/index.tsx
@@ -3,9 +3,9 @@ import { GetStaticProps, NextPage } from "next";
 import { useSnackbar } from "notistack";
 import { useState } from "react";
 import useSWR from "swr";
-import AddEventDialog from "../../../components/AddEventDialog";
+import AddEventDialog from "../../../components/admin/AddEventDialog";
+import EventList from "../../../components/admin/EventList";
 import { useConfirm } from "../../../components/ConfirmProvider";
-import EventList from "../../../components/EventList";
 import AdminLayout from "../../../components/layouts/AdminLayout";
 import { Event, hasPermission, User } from "../../../interfaces";
 import { AuthContextInstance, withAuth } from "../../../lib/auth";
