+++ src/pages/admin/users/[id].tsx
@@ -2,9 +2,9 @@ import { Container, Typography } from "@material-ui/core";
 import { makeStyles, Theme } from "@material-ui/core/styles";
 import { GetStaticPaths, GetStaticProps, GetStaticPropsResult, NextPage } from "next";
 import { useRouter } from "next/router";
-import AdminLoading from "../../../components/AdminLoading";
+import AdminLoading from "../../../components/admin/AdminLoading";
+import ListDetail from "../../../components/admin/ListDetail";
 import AdminLayout from "../../../components/layouts/AdminLayout";
-import ListDetail from "../../../components/ListDetail";
 import { hasPermission, User } from "../../../interfaces";
 import { AuthContextInstance, ExcludeAuthProps, withAuth } from "../../../lib/auth";
 import { getUser } from "../../../lib/db";
