+++ src/pages/admin/users/index.tsx
@@ -2,8 +2,8 @@ import { Container, makeStyles, Theme, Typography } from "@material-ui/core";
 import { GetStaticProps, NextPage } from "next";
 import { useSnackbar } from "notistack";
 import useSWR from "swr";
+import UserList from "../../../components/admin/UserList";
 import AdminLayout from "../../../components/layouts/AdminLayout";
-import UserList from "../../../components/UserList";
 import { hasPermission, User } from "../../../interfaces";
 import { AuthContextInstance, withAuth } from "../../../lib/auth";
 import { getAllUsers } from "../../../lib/db";
